Now the toy simulation takes into account the simulation volume
The fraction of mass of the simulation is now calculated through the
formula r*exp(r/rd)*dr/(rd*rd). This formula comes from the total
integration of the mass of the galaxy (rd*rd) and the "linear"
approximation to the integral: dM = r*exp(r/rd)*dr.

The input example file has been modified accordingly.