Fixed the apparent structure that appeared in some runs
The key was the rounding up of the number of events to get an integer
for the random function. If the change between one calculation of the
event routine and the next was in the single digits, it would
appreciably appear as structure in the final run. The solution is to
dynamically increase the SFR sampling so that the variation between one
step and the next is big enough to not be rounded away when converting
to integer.

Also changed the visualizeOutput.py script to compute the ratio of two
elements if two inputs are added.